requirements.txt
requests
ipaddress
dnspython
re
os
time 
datetime